requirements.txt
absl-py==2.1.0
accelerate==1.8.1
albucore==0.0.33
albumentationsx==2.0.9
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==25.1.0
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
efficientnet_pytorch==0.7.1
einops==0.8.1
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.2.0
grpcio==1.67.0
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.33.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
kornia==0.8.0
kornia_rs==0.1.9
lazy_loader==0.4
lightning-utilities==0.15.2
llvmlite==0.44.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
MedPy==0.5.2
mistune==3.0.2
mpmath==1.3.0
mrcfile==1.5.4
munch==4.0.0
mypy_extensions==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
onnx==1.17.0
onnxruntime-gpu==1.20.1
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patchify==0.2.3
pathspec==0.12.1
pexpect==4.9.0
phasepack==1.5
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
pretrainedmodels==0.7.4
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyFFTW==0.15.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch_toolbelt==0.8.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ridge-detection==3.0.0
ridge-detector==0.1.2
rpds-py==0.20.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
segment-anything==1.0
segmentation_models_pytorch==0.5.0
Send2Trash==1.8.3
simpleitk==2.5.2
simplejson==3.20.1
simsimd==6.5.0
six==1.16.0
skorch==1.1.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
stringzilla==3.12.5
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.19
timm-3d==1.0.1
tinycss2==1.3.0
tokenizers==0.21.2
torch==2.7.1+cu118
torch-tb-profiler==0.4.3
torchaudio==2.7.1+cu118
torchmetrics==1.8.1
torchvision==0.22.1+cu118
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.53.1
triton==3.3.1
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
yarg==0.1.9
